Support list of architectures for Debian packages

+++ deploy/debian/package/package.sh
@@ -44,9 +44,15 @@ ${PKG_NAME} ($VERSION) unstable; urgency=low
 License: ${PKG_LICENSE}" > ${DEBIAN_DIR}/copyright
 
     docker build -t rygel/${DOCKER_IMAGE} deploy/docker/${DOCKER_IMAGE}
-    docker run --privileged -t -i --rm -v $(pwd):/io/host -v $(pwd)/${CLIENT_DIR}:/io/client rygel/${DOCKER_IMAGE} /io/host/${SCRIPT_PATH} build x86_64-linux-gnu amd64 x64
-    docker run --privileged -t -i --rm -v $(pwd):/io/host -v $(pwd)/${CLIENT_DIR}:/io/client rygel/${DOCKER_IMAGE} /io/host/${SCRIPT_PATH} build i686-linux-gnu i386 x86
-    docker run --privileged -t -i --rm -v $(pwd):/io/host -v $(pwd)/${CLIENT_DIR}:/io/client rygel/${DOCKER_IMAGE} /io/host/${SCRIPT_PATH} build aarch64-linux-gnu arm64 ARM64
+    if echo "${PKG_ARCHITECTURES}" | grep -q -w amd64; then
+        docker run --privileged -t -i --rm -v $(pwd):/io/host -v $(pwd)/${CLIENT_DIR}:/io/client rygel/${DOCKER_IMAGE} /io/host/${SCRIPT_PATH} build x86_64-linux-gnu amd64 x64
+    fi
+    if echo "${PKG_ARCHITECTURES}" | grep -q -w i386; then
+        docker run --privileged -t -i --rm -v $(pwd):/io/host -v $(pwd)/${CLIENT_DIR}:/io/client rygel/${DOCKER_IMAGE} /io/host/${SCRIPT_PATH} build i686-linux-gnu i386 x86
+    fi
+    if echo "${PKG_ARCHITECTURES}" | grep -q -w arm64; then
+        docker run --privileged -t -i --rm -v $(pwd):/io/host -v $(pwd)/${CLIENT_DIR}:/io/client rygel/${DOCKER_IMAGE} /io/host/${SCRIPT_PATH} build aarch64-linux-gnu arm64 ARM64
+    fi
 
     cp ${CLIENT_DIR}/upper/*/${DEST_DIR}/${PKG_NAME}_*.deb ${PKG_DIR}/
 elif [ "$1" = "build" ]; then
